Laravel 5.8 renders throttle security check useless
As of Laravel 5.8 the cache expiration time changed from minutes to seconds.  This means this security check does not work any longer as the cache expires after 1 second rendering it useless.  This could be fixed by simply multiplying all of the settings values by 60 to get number of seconds but it may break again in the future.  The proper fix is to pass in a valid Carbon date for expiration instructing the cache to expire at an exact time.